requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.1
cryptography==35.0.0
Deprecated==1.2.13
Flask==2.0.2
gcloud==0.18.3
googleapis-common-protos==1.53.0
httplib2==0.20.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
jwcrypto==1.0
MarkupSafe==2.0.1
oauth2client==4.1.3
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.11.0
pyparsing==2.4.7
Pyrebase4==4.5.0
python-jwt==3.3.0
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.16.0
urllib3==1.26.7
uuid==1.30
Werkzeug==2.0.2
wrapt==1.13.2
